total price counter
Added a helper function that calculates the total price for the entire cart by iterating over all items in the cartItems array using the reduce function. For each item, it multiplies item.product.price by item.quantity to get the total price for that specific item, then adds up the total price for all items and returns the final sum. Finally, I implemented the helper function in both addProductToCart and removeCartItem to update the state(totalPrice).